channellist logic to chat.js
